CSS as file.
Extracts the inline styles from style.xsls into separate files for
ease of style-development and to take advantage of browser caching.
Hebrew and Chinese variants are preserved.

Stylesheets are minified as part of make process and served from css
subdirectory. The minified CSS is compressed during the gzip process. A
Perl script (located in new tools subdirectory) performs the minification.