Enabling local development.
This commit enables the nginx.org repository for external contributions and
makes it easier for others to build and develop the website locally.

Now licensed under CC4.0.

A new tools directory now houses the scripts necessary to build the site.
Note that xslscript.pl was previously hosted in a separate repo and was
copied from <https://hg.nginx.org/xslscript/>.

Dockerfile creates a self-contained image that can be used for staging
and/or hosting of the website.

docker-compose.yml facilitates local development by running an nginx image
configured to serve from the local repository.

docker-nginx.conf provides a minimal NGINX configuration for the Docker
environments.